Add max_plain_text_length option to ssl:listen.
This corresponds to OpenSSL's SSL_CTRL_SET_MAX_SEND_FRAGMENT option, and
allows the size of plain text fragments to be specified.

This is useful, for example, for talking to embedded devices with
limited resources.

This is different from the max_fragment_length negotiation specified by
RFC 6066:

- RFC 6066 is an extension to TLS 1.2; this supports TLS 1.1.
- RFC 6066 only allows 512-, 1024-, 2048- and 4096-byte fragment sizes;
  this allows arbitrary values, 512 <= N <= 16384.

Unfortunately, this requires the server to _know_ what the client's
expecting; this is an acceptable trade-off in this scenario.